Replace CatalogController#index and SolrHelper#get_single_doc_via_search over-rides with a single SolrHelper#solr_search_params over-ride.
previously, the plugin tried to over-ride (via MonkeyPatch) CatalogController#index, and SolrHelper#get_single_doc_via_search  . Note that SolrHelper is a module of CatalogController already.

The functionality in both of these places can be dry'd up simply by over-riding SolrHelper#solr_search_params instead -- if we do this as an actual ruby include instead of relying on Engine auto-monkey-patching, we are safer in terms of relying on order of loading, AND we can call super in our implementation to not copy-and-paste any code from Blacklight core, safer for future BL changes.