Change controller_override to use new BL 2.9+ solr_search_params_logic
Allows for somewhat easier to debug and clearer code to add in logic
to solr_search_params, in this case for advanced search parsing.

Will make it easier to add an additional feature to adv search,
ability to have 'basic search' be parsed for AND/OR/NOT too.

Required changing of solr params hash merging, under new design we
WANT to actually mutate the original hash, rather than wanting to make
sure not to mutate it. However, we stil want to use custom merge logic,
so array keys get merged.