REMOVE solr_type config to choose parsing strategy
* 'edismax' was unfinished/abandoned, and it's time to trim it.
* 'dismax' is superceded by nesting_parsing, too hard to support both
   when adding new features, like the new advanced_parse_q

Removing both of those, no need for solr_type config anymore. That
was a bad name anyway, it ended up describing the plugin's parse
strategy/method more than it did "solr type". if we add alternate strategies
again in the future, we will use a different name for the config.