more careful context search
We're trying to find existing facets within just a PART of the
current search. May be too weird a thing we're trying to do
in the first place, but that's what its' been. Make sure
to exclude adv search criteria, we want existing facets
within current search not included adv search criteria and
'q'