generator uses actual BL source template
Install generator will optionally copy a _search_form template into
your local app, with a link to advanced search. Previously it used
it's own local template for search_form. But then this had to be sync'd
with whatever was in current Blacklight.

Changed to actually read app/views/catalog/_search_form from currently
loaded Blacklight gem with `Blacklight.root`, add a link to advanced
search form to end of it, and copy that into local app. This way
it will always use appropriate search form for currently loaded BL.

If you later upgrade BL, you'll still have to look at if search_form
has changed and merge changes in, but that's how it goes with localized
templates. If that's a problem, we'd have to add a feature to BL to have
a hook for third party code to insert things into end of search_form
template.