proper interaction of explicit advanced 'qt' and qt from host BL
cbeer pointed out we can just let solr_search_params set ordinary qt and not try to duplicate.
looks like we were erreneously defaulting to an explicit qt of 'advanced' when README
said otherwise; not sure if it had that effect previously, maybe existing code worked
around it, not sure.