Rescue Parslet parsing errors on advanced search ...
... and raise Blacklight's InvalidRequest error.

This currently happens when processing queries in for advanced search features in the single-search-box/q param, but not for advanced search itself.

This will cause the error if only one field has a parsing error, but probably a good idea to indicate that to the user and allow them to hit the back button and fix their search (as opposed to getting a generic rails error page).